new file to handle the penglai ocall
