+++ enclave-driver/penglai-enclave-ioctl.c
@@ -3,6 +3,7 @@
 #include "penglai-shm.h"
 #include "penglai-schrodinger.h"
 #include "penglai-enclave-persistency.h"
+#include "penglai-enclave-ocall.h"
 #include <linux/delay.h>
 
 //TODO: improve concurrency
@@ -21,7 +22,7 @@ int create_sbi_param(enclave_t* enclave, struct penglai_enclave_sbi_param * encl
     unsigned long paddr, unsigned long size, unsigned long entry_point,
     unsigned long free_mem,
     unsigned long shm_vaddr, unsigned long shm_size)
-{
+{ 
   enclave_sbi_param->eid_ptr = (unsigned int* )__pa(&enclave->eid);
   enclave_sbi_param->ecall_arg0 = (unsigned long* )__pa(&enclave->ocall_func_id);
   enclave_sbi_param->ecall_arg1 = (unsigned long* )__pa(&enclave->ocall_arg0);
@@ -305,144 +306,6 @@ int penglai_enclave_resume_for_rerun(enclave_instance_t *enclave_instance, encla
     {
       penglai_eprintf("penglai_enclave_resume_for_rerun: ocall func id is error\n");
       ret = SBI_PENGLAI_2(SBI_SM_RESUME_ENCLAVE, resume_id, RESUME_FROM_OCALL);
-    }
-  }
-  return ret;
-}
-
-int penglai_enclave_ocall(enclave_instance_t *enclave_instance, enclave_t *enclave, int resume_id, int isShadow)
-{
-  unsigned int ocall_func_id;
-  int ret;
-  ocall_func_id = enclave->ocall_func_id;
-  if(isShadow)
-    ocall_func_id = enclave_instance->ocall_func_id;
-  switch(ocall_func_id)
-  {
-    case OCALL_MMAP:
-    {
-      uintptr_t order, vaddr;
-      if(isShadow)
-        order = ilog2((enclave_instance->ocall_arg1 >> RISCV_PGSHIFT) - 1) + 1;
-      else
-        order = ilog2((enclave->ocall_arg1 >> RISCV_PGSHIFT) - 1) + 1;
-      vaddr = penglai_get_free_pages(GFP_KERNEL, order);
-      if(!vaddr)
-      {
-        ret = -1;
-        penglai_eprintf("penglai_enclave_ocall: OCALL_MMAP  is failed\r\n");
-        break;
-      }
-      ret = SBI_PENGLAI_5(SBI_SM_RESUME_ENCLAVE, resume_id, RESUME_FROM_OCALL, OCALL_MMAP, __pa(vaddr), (1<<order)*RISCV_PGSIZE);
-      break;
-    }
-    case OCALL_UNMAP:
-    {
-      uintptr_t vaddr, order;
-      if(isShadow)
-      {
-        vaddr = (uintptr_t)__va(enclave_instance->ocall_arg0);
-        order = ilog2((enclave_instance->ocall_arg1 >> RISCV_PGSHIFT) - 1) + 1;
-      }
-      else
-      {
-        vaddr = (uintptr_t)__va(enclave->ocall_arg0);
-        order = ilog2((enclave->ocall_arg1 >> RISCV_PGSHIFT) - 1) + 1;
-      }
-      
-      free_pages(vaddr, order);
-      ret = SBI_PENGLAI_3(SBI_SM_RESUME_ENCLAVE, resume_id, RESUME_FROM_OCALL, OCALL_UNMAP);
-      break;
-    }
-    case OCALL_SYS_WRITE:
-    {
-      if(isShadow)
-      {
-        ((char*)(enclave_instance->kbuffer))[511] = '\0';
-        printk((void*)(enclave_instance->kbuffer));
-      }
-      else
-      {
-        ((char*)(enclave->kbuffer))[511] = '\0';
-        printk((void*)(enclave->kbuffer));
-      }
-      ret = SBI_PENGLAI_3(SBI_SM_RESUME_ENCLAVE, resume_id, RESUME_FROM_OCALL, OCALL_SYS_WRITE);
-      break;
-    }
-    case OCALL_SBRK:
-    {
-      long size;
-      uintptr_t vaddr = 0;
-      uintptr_t order = 0;
-      struct pm_area_struct *pma;
-      if(isShadow)
-      {
-        size = (long)(enclave_instance->ocall_arg1);
-        pma = (struct pm_area_struct*)(enclave_instance->ocall_arg0);
-      }
-      else
-      {
-        pma = (struct pm_area_struct*)(enclave->ocall_arg0);
-        size = (long)(enclave->ocall_arg1);
-      }
-      
-      if(size < 0)
-      {
-        while(pma)
-        {
-          pma = (struct pm_area_struct*)__va(pma);
-          vaddr = (uintptr_t)__va(pma->paddr);
-
-          order = ilog2((pma->size >> RISCV_PGSHIFT) - 1) + 1;
-          pma = pma->pm_next;
-          //should be freed after set pma to its next
-          free_pages(vaddr, order);
-        }
-      }
-      else
-      {
-        order = ilog2((size >> RISCV_PGSHIFT) - 1) + 1;
-        vaddr = penglai_get_free_pages(GFP_KERNEL, order);
-        if(!vaddr)
-        {
-          penglai_eprintf("penglai_enclave_ocall: OCALL_SBRK is failed\r\n"); 
-          ret = -1;
-          break;
-        }
-      }
-      ret = SBI_PENGLAI_5(SBI_SM_RESUME_ENCLAVE, resume_id, RESUME_FROM_OCALL, OCALL_SBRK, __pa(vaddr), (1<<order)*RISCV_PGSIZE);
-      break;
-    }
-    default:
-    {
-      if(ocall_func_id == OCALL_READ_SECT)
-      {
-        int read;
-        if (isShadow)
-          read = penglai_outer_read(enclave_instance->ocall_arg0);
-        else
-          read = penglai_outer_read(enclave->ocall_arg0);
-        
-        ret = SBI_PENGLAI_4(SBI_SM_RESUME_ENCLAVE, resume_id,RESUME_FROM_OCALL, OCALL_READ_SECT, read);
-      }
-      else if(ocall_func_id == OCALL_WRITE_SECT)
-      {
-        int write;
-        if (isShadow)
-        write = penglai_outer_write(enclave_instance->ocall_arg0);
-        else
-          write = penglai_outer_write(enclave->ocall_arg0);
-        
-        ret = SBI_PENGLAI_4(SBI_SM_RESUME_ENCLAVE, resume_id, RESUME_FROM_OCALL, OCALL_WRITE_SECT,write);
-      }
-      else if (ocall_func_id == OCALL_RETURN_RELAY_PAGE)
-      {
-        ret = ENCLAVE_RETURN_USER_MODE;
-        break;
-      }
-      else{
-        ret = SBI_PENGLAI_2(SBI_SM_RESUME_ENCLAVE, resume_id, RESUME_FROM_OCALL);
-      }
     }
   }
   return ret;
